requirements.txt
photoshop_python_api==0.21.1
openpyxl==3.1.5